requirements.txt
python-telegram-bot[rate-limiter,callback-data]==20.7
pyyaml>=6.0
cryptography>=41.0
uvloop>=0.17; sys_platform != 'win32'
aiofiles>=23.1
pytest>=7.4
fastapi>=0.110
uvicorn>=0.23
